Implement 2-stage commit for CEvoDB to avoid inconsistencies
>>> from dash@521d4ae08fe0e35c30bf05ffe8b9d611e86c3bfd

CDBTransaction is changed to allow CDBBatch, CDBWrapper and other
CDBTransactions as parent instead of just CDBWrapper. This in turn
allows to implement multi-staged commits in CEvoDB.

We now have the "current transaction" which is started and ended (commit
or rollback) for each call to Connect-/DisconnectBlock. When the current
transaction is committed, it moves its contents into the "root
transaction" instead of directly writing to CDBWrapper.
CommitRootTransaction() then handles the final commitment to CDBWrapper.
It is called at the same time when the chainstate is flushed to disk,
which guarantees consistency between chainstate and CEvoDB.