[test] Add CCoinsViewCache Access/Modify/Write tests
>>> backports bitcoin/bitcoin@07df40babb0231b2e8a2f42341805b4bbcc4e284

Add more comprehensive unit tests for CCoinsViewCache. Right now it is
hard to refactor caching code or fix bugs in the caching logic because
you have to try to mentally enumerate all the different states the cache
might be in to make sure a change doesn't cause unintended consequences.
The new tests explicitly enumerate relevant cache states, documenting
and verifying the behavior in each state, so it will be safer and easier
to make changes to the caching code in the future.