Eliminate TX trickle bypass, sort TX invs for privacy and priority.
Previously Bitcoin would send 1/4 of transactions out to all peers
instantly. This causes high overhead because it makes >80% of
INVs size 1. Doing so harms privacy, because it limits the
amount of source obscurity a transaction can receive.

These randomized broadcasts also disobeyed transaction dependencies
and required use of the orphan pool. Because the orphan pool is
so small this leads to poor propagation for dependent transactions.

When the bypass wasn't in effect, transactions were sent in the
order they were received. This avoided creating orphans but
undermines privacy fairly significantly.

This commit:
Eliminates the bypass. The bypass is replaced by halving the
 average delay for outbound peers.

Sorts candidate transactions for INV by their topological
 depth then by their feerate (then hash); removing the
 information leakage and providing priority service to
 higher fee transactions.

Limits the amount of transactions sent in a single INV to
 7tx/sec (and twice that for outbound); this limits the
 harm of low fee transaction floods, gives faster relay
 service to higher fee transactions. The 7 sounds lower
 than it really is because received advertisements need
 not be sent, and because the aggregate rate is multipled
 by the number of peers.

 Coming from btc@f2d3ba73860e875972738d1da1507124d0971ae5