[BUG] Fix mempool conflicts due to disconnected blocks
If, during a reorg, a ProReg tx is disconnected, and not reconnected in
the other chain, we need to evict any mempool transaction
that depends on the masternode registered by the ProReg, before adding
it back to the mempool.

Otherwise (as the tiertwo_reorg_mempool functional test shows) a node,
after a reorganization, could try to mine a block which includes a
ProUpServ transaction linked to a no-longer-existing masternode.